pyproject.toml
[tool.poetry]
name = "graph-semantic-search"
version = "0.1.0"
description = ""
authors = ["Andy Chumak <andrii.chumak@gooddata.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
neo4j = "^5.20.0"
python-dotenv = "^1.0.1"
gooddata-sdk = "^1.18.1"
transformers = "^4.40.1"
torch = "^2.3.0"
gradio = "^4.28.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
